修改代码生成代码的读取环境变量方式

+++ database/src/main/java/com/zanghongtu/database/Dao2POJO.java
@@ -4,43 +4,31 @@ package com.zanghongtu.database;
 import com.zanghongtu.alphabeta.common.CamelCase;
 import com.zanghongtu.alphabeta.common.file.FileOperator;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
+import javax.naming.Context;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.lang.annotation.Annotation;
+import java.net.URLDecoder;
 import java.sql.*;
-import java.util.HashSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
 
 /**
  * @author hongtu
  */
-@Component
+@Service
 public class Dao2POJO {
-    @Value("${database.package}")
-    private String basePackage;
-
-    @Value("${database.service}")
-    private Boolean initServices;
-
-    @Value("${database.repository}")
-    private Boolean initRepositories;
-
-    @Value("${database.model}")
-    private Boolean initModels;
-
-    @Value("${spring.datasource.url}")
     private String jdbcUrl;
 
-    @Value("${spring.datasource.username}")
     private String dbUser;
 
-    @Value("${spring.datasource.password}")
     private String dbPwd;
 
     private String srcBasePath;
@@ -55,7 +43,14 @@ public class Dao2POJO {
         defaultColumns.add("available");
     }
 
+    public void init(String jdbcUrl, String dbUser, String dbPwd) {
+        this.jdbcUrl = jdbcUrl;
+        this.dbUser = dbUser;
+        this.dbPwd = dbPwd;
+    }
+
     public void generateCodes() {
+        String basePackage = findSpringBootApplicationPackage();
         srcBasePath = System.getProperty("user.dir") + "/src/main/java/" + basePackage.replaceAll("\\.", "/") + "/";
         String dbName = jdbcUrl.split("[/?]")[3];
         try (Connection connection = DriverManager.getConnection(jdbcUrl, dbUser, dbPwd)) {
@@ -69,19 +64,16 @@ public class Dao2POJO {
             //获取表的字段，拼成POJO
             for (String tableName : tableNames) {
                 String className = CamelCase.toCamelCaseClassName(tableName);
-                writePojos(connection, tableName, className);
-                writeRepositories(className);
-                writeServices(className);
+                writePojos(connection, tableName, className, basePackage);
+                writeRepositories(className, basePackage);
+                writeServices(className, basePackage);
             }
         } catch (SQLException e) {
             e.printStackTrace();
         }
     }
 
-    private void writePojos(Connection connection, String tableName, String className) {
-        if (!initModels) {
-            return;
-        }
+    private void writePojos(Connection connection, String tableName, String className, String basePackage) {
         String sql = "SELECT * FROM " + tableName + " WHERE 1=2;";
         try (PreparedStatement ps = connection.prepareStatement(sql);
              ResultSet rs = ps.executeQuery()) {
@@ -100,7 +92,7 @@ public class Dao2POJO {
                 columnStr.append("    private ").append(colType).append(" ").append(paramName).append(";\n\n");
             }
             String content = readTemplate("model")
-                    .replaceAll("###BASE_PACKAGE###", this.basePackage)
+                    .replaceAll("###BASE_PACKAGE###", basePackage)
                     .replaceAll("###CLASS_NAME###", className)
                     .replaceAll("###COLUMNS###", columnStr.toString())
                     .replaceAll("###TABLE_NAME###", tableName);
@@ -111,39 +103,47 @@ public class Dao2POJO {
         }
     }
 
-    private void writeServices(String className) {
-        if (!initServices) {
-            return;
-        }
-        writeServiceCode(className);
-        writeServiceImplCode(className);
+    private void writeServices(String className, String basePackage) {
+        writeServiceCode(className, basePackage);
+        writeServiceImplCode(className, basePackage);
     }
 
-    private void writeServiceImplCode(String className) {
+    private void writeServiceImplCode(String className, String basePackage) {
         String content = readTemplate("serviceImpl")
-                .replaceAll("###BASE_PACKAGE###", this.basePackage)
+                .replaceAll("###BASE_PACKAGE###", basePackage)
                 .replaceAll("###CLASS_NAME###", className);
         String filePath = srcBasePath + "service/impl/" + className + "ServiceImpl.java";
-        FileOperator.writeFile(filePath, content);
-    }
-
-    private void writeServiceCode(String className) {
-        String content = readTemplate("service")
-                .replaceAll("###BASE_PACKAGE###", this.basePackage)
-                .replaceAll("###CLASS_NAME###", className);
-        String filePath = srcBasePath + "service/" + className + "Service.java";
-        FileOperator.writeFile(filePath, content);
-    }
-
-    private void writeRepositories(String className) {
-        if (!initRepositories) {
+        if (exist(filePath)) {
             return;
         }
+        FileOperator.writeFile(filePath, content);
+    }
+
+    private void writeServiceCode(String className, String basePackage) {
+        String content = readTemplate("service")
+                .replaceAll("###BASE_PACKAGE###", basePackage)
+                .replaceAll("###CLASS_NAME###", className);
+        String filePath = srcBasePath + "service/" + className + "Service.java";
+        if (exist(filePath)) {
+            return;
+        }
+        FileOperator.writeFile(filePath, content);
+    }
+
+    private void writeRepositories(String className, String basePackage) {
         String content = readTemplate("repository")
-                .replaceAll("###BASE_PACKAGE###", this.basePackage)
+                .replaceAll("###BASE_PACKAGE###", basePackage)
                 .replaceAll("###CLASS_NAME###", className);
         String filePath = srcBasePath + "repository/" + className + "Repository.java";
+        if (exist(filePath)) {
+            return;
+        }
         FileOperator.writeFile(filePath, content);
+    }
+
+    private boolean exist(String filePath) {
+        File file = new File(filePath);
+        return file.exists();
     }
 
     private String readTemplate(String templateName) {
@@ -164,6 +164,92 @@ public class Dao2POJO {
             e.printStackTrace();
         }
         return null;
+    }
+
+    private String findSpringBootApplicationPackage() {
+        String url = System.getProperty("user.dir");
+        List<String> classes = getClassesList(url);
+        // 遍历classes，如果发现@Component就注入到容器中
+        return scanSpringBootApplication(classes);
+    }
+
+    private String scanSpringBootApplication(List<String> classes) {
+        for (String className : classes) {
+            Class clazz = null;
+            try {
+                clazz = Class.forName(className);
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            }
+            Annotation[] annotations = new Annotation[0];
+            if (clazz != null) {
+                annotations = clazz.getDeclaredAnnotations();
+            }
+            for (Annotation annotation : annotations) {
+                if (annotation instanceof SpringBootApplication) {
+                    return clazz.getPackage().getName();
+                }
+            }
+        }
+        return null;
+    }
+
+    private String getClassPath() {
+        String url = null;
+        try {
+            url = URLDecoder.decode(Context.class.getResource("/").getPath(), "utf-8");
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
+        }
+        if (url != null && url.startsWith("/")) {
+            url = url.replaceFirst("/", "");
+            url = url.replaceAll("/", "\\\\");
+        }
+        return url;
+    }
+
+    private List<String> getClassesList(String url) {
+        File file = new File(url);
+        List<String> classes = getAllClass(file);
+        for (int i = 0; i < classes.size(); i++) {
+            classes.set(i, classes.get(i).replace(url, "").replace(".java", "")
+                    .replace("\\", "."));
+        }
+        List<String> result = new LinkedList<>();
+        for (String classFile : classes) {
+            if (!classFile.startsWith("/src/main/java")) {
+                continue;
+            }
+            result.add(classFile.replace("/src/main/java/", "")
+                    .replace("/", "."));
+        }
+        return result;
+    }
+
+    private List<String> getAllClass(File file) {
+        List<String> ret = new ArrayList<>();
+        if (file.isDirectory()) {
+            File[] list = file.listFiles();
+            if (list != null) {
+                for (File i : list) {
+                    List<String> j = getAllClass(i);
+                    ret.addAll(j);
+                }
+            }
+        } else {
+            ret.add(file.getAbsolutePath());
+        }
+        return ret;
+    }
+
+    private void fetchFileList(File dir, List<File> fileList) {
+        if (dir.isDirectory()) {
+            for (File f : Objects.requireNonNull(dir.listFiles())) {
+                fetchFileList(f, fileList);
+            }
+        } else {
+            fileList.add(dir);
+        }
     }
 }
